관리자 유저 구분용 isAdmin 추가

+++ src/main/java/com/gajimarket/Gajimarket/admin/AdminController.java
@@ -21,28 +21,45 @@ public class AdminController {
 
     // 전체 사용자 정보 가져오기
     @GetMapping("/get/userList")
-    public ResponseEntity<ApiResponse> getUserList(){
+    public ResponseEntity<ApiResponse> getUserList(@RequestParam boolean isAdmin){
         System.out.println("/get/userList request");
+
         ApiResponse apiResponse;
-        try{
-            //서비스 호출
-            List<UserInfo> userList=adminService.getUserList();
-            apiResponse=new ApiResponse("1000", null);
+
+        try {
+            // isAdmin이 false인 경우 관리자 권한이 없으므로 예외 처리
+            if (!isAdmin) {
+                apiResponse = new ApiResponse("500", "관리자 권한이 없음");
+                return ResponseEntity.ok().body(apiResponse);
+            }
+
+            // 서비스 호출
+            List<UserInfo> userList = adminService.getUserList();
+            apiResponse = new ApiResponse("1000", null);
             apiResponse.setData(userList);
             return ResponseEntity.ok().body(apiResponse);
 
-        } catch (Exception e){
+        } catch (Exception e) {
             apiResponse = new ApiResponse("0", "리스트 불러오기 실패");
             return ResponseEntity.ok().body(apiResponse);
         }
     }
 
+
     // 회원 영구 정지(매너 지수 -1로 설정)
     @PatchMapping("/ban/{userIdx}")
-    public ResponseEntity<ApiResponse> banUser(@PathVariable("userIdx") int userIdx) {
+    public ResponseEntity<ApiResponse> banUser(
+            @RequestParam boolean isAdmin,
+            @PathVariable("userIdx") int userIdx) {
         System.out.println("/admin/ban/" + userIdx + " request");
         ApiResponse apiResponse;
         try {
+            // isAdmin이 false인 경우 관리자 권한이 없으므로 예외 처리
+            if (!isAdmin) {
+                apiResponse = new ApiResponse("500", "관리자 권한이 없음");
+                return ResponseEntity.ok().body(apiResponse);
+            }
+
             adminService.banUser(userIdx);
             apiResponse = new ApiResponse("1000", "회원 영구 정지 성공");
             return ResponseEntity.ok().body(apiResponse);
@@ -53,11 +70,16 @@ public class AdminController {
     }
     // 관리자 목록 가져오기
     @GetMapping("/get/adminList")
-    public ResponseEntity<ApiResponse> getAdminList() {
+    public ResponseEntity<ApiResponse> getAdminList(@RequestParam boolean isAdmin) {
         System.out.println("/admin/get/adminList request");
 
         ApiResponse apiResponse;
         try {
+            // isAdmin이 false인 경우 관리자 권한이 없으므로 예외 처리
+            if (!isAdmin) {
+                apiResponse = new ApiResponse("500", "관리자 권한이 없음");
+                return ResponseEntity.ok().body(apiResponse);
+            }
             // 서비스 호출
             List<UserInfo> adminList = adminService.getAdminList();
             apiResponse = new ApiResponse("1000", null);
@@ -71,11 +93,18 @@ public class AdminController {
     }
 
     @PostMapping("/add")
-    public ResponseEntity<ApiResponse> addAdmin(@RequestBody AddAdminRequest request) {
+    public ResponseEntity<ApiResponse> addAdmin(
+            @RequestParam boolean isAdmin,
+            @RequestBody AddAdminRequest request) {
         System.out.println("/admin/add request");
 
         ApiResponse apiResponse;
         try {
+            // isAdmin이 false인 경우 관리자 권한이 없으므로 예외 처리
+            if (!isAdmin) {
+                apiResponse = new ApiResponse("500", "관리자 권한이 없음");
+                return ResponseEntity.ok().body(apiResponse);
+            }
             // 서비스 호출
             adminService.addAdmin(request);
             apiResponse = new ApiResponse("1000", "관리자 추가 성공");
@@ -93,11 +122,18 @@ public class AdminController {
     }
 
     @DeleteMapping("/delete/{admin_idx}")
-    public ResponseEntity<ApiResponse> deleteAdmin(@PathVariable int admin_idx) {
+    public ResponseEntity<ApiResponse> deleteAdmin(
+            @RequestParam boolean isAdmin,
+            @PathVariable int admin_idx) {
         System.out.println("/admin/delete/" + admin_idx + " request");
         ApiResponse apiResponse;
 
         try {
+            // isAdmin이 false인 경우 관리자 권한이 없으므로 예외 처리
+            if (!isAdmin) {
+                apiResponse = new ApiResponse("500", "관리자 권한이 없음");
+                return ResponseEntity.ok().body(apiResponse);
+            }
             // 서비스 호출
             adminService.deleteAdmin(admin_idx);
             apiResponse = new ApiResponse("1000", "관리자 삭제 성공");
@@ -123,10 +159,15 @@ public class AdminController {
 
     // 신고 목록 가져오기
     @GetMapping("/get/reportList")
-    public ResponseEntity<ApiResponse> getReportList(){
+    public ResponseEntity<ApiResponse> getReportList(@RequestParam boolean isAdmin){
         System.out.println("/get/reportList request");
         ApiResponse apiResponse;
         try{
+            // isAdmin이 false인 경우 관리자 권한이 없으므로 예외 처리
+            if (!isAdmin) {
+                apiResponse = new ApiResponse("500", "관리자 권한이 없음");
+                return ResponseEntity.ok().body(apiResponse);
+            }
             //서비스 호출
             List<ReportInfo> reportList = adminService.getReportList();
             apiResponse=new ApiResponse("1000", null);
